Added conditional cartpole and improve argument logic for condition types
