+++ .gitignore
@@ -139,3 +139,4 @@ iq_learn/data
 *.png
 *.zip
 *.txt
+*.pkl
